Implement frame search
Adapted the iterator protocol so it returns `Option<FrameResult>`
instead of `Option<Frame>` now. This means we can have an iterator that
returns `Some(FrameResult::Fail)` for frames that fail to meet the
filter, and we can take advantage of early-returning instead of
generating the entire frame before checking if it passes, but we don't
have to end the iterator as soon as the first frame fails.