Pass seed as two u32s instead of one u64
For some reason wasm-opt can't optimize the build if any function
returns a u64 or exposes a pub u64 struct field. We need to get around
this to pass the seed for each frame, which is a u64. So we split it
into two u32 components, and pass them as a tuple struct. Then we
can use bitwise operations on BigInts to reconstitute the seed in
JavaScript.