Fix Parcel type imports
The previous fix allowed TypeScript to import from
`../crate/Cargo.toml`, with two downsides:

1. The module was defined as having a single default export, which
disallowed `import { A, B }` or `import type { A, B }`
2. Only terms were visible (enums, functions), but not types, since
the import was not recognized as a namespace

The new declaration fixes the problem:

1. We still use dynamic `import()` since this is somehow necessary
to load types
2. `export * from "../crate/pkg/raidtomi"` instead of `export default`

In addition, we can remove the `<reference path>` shenanigans by
renaming `index.d.ts` to `crate.d.ts`. Having multiple `index.*` files
was confusing `tsc`. Thanks to cspotcode on the TypeScript Discord.