Reset gender filter when encounter changes
Different Pokemon have different gender ratios and gender locks, which
affects the set of valid gender filters.

When changing the den or encounter, the mon's gender stats might change,
so we reset the gender filter to `undefined`. We do the same when
updating settings that transitively change the encounter, e.g. the game
title or badge level.

In almost every case this is what we want. If the mon or raid is
gender-locked, the frame generation process will always produce the right
gender, so specifying a gender is unnecessary.

- Locked -> Locked: undefined -> undefined
- Locked -> Random: undefined -> undefined (new gender pool anyway)
- Random -> Locked: possibly defined -> undefined (old filter doesn't
matter anyway)
- Random -> Random: this is the only case that can be annoying.

Let's say you're hunting for a female Ralts, and you switch to Kirlia in
the same den. Ideally in this case if the filter is valid for the new
mon we keep it. But this would require lifting up the legality logic
to the top-level component, which I don't want to deal with now.