Implement NatureFilter::from_natures
Due to wasm-bindgen's limitations [1], we need a constructor that
takes a `Vec<u32>` since we cannot take a `Vec<Nature>`. As a result, in
Rust-land we go through annoying acrobatics to convert a `Vec<Nature>`
into a `Vec<u32>` everywhere.

Instead, we simply implement a `NatureFilter::from_natures` method that
is not exposed to JS. This allows our tests to pass a `Vec<Nature>`,
which is much more ergonomic.

[1] https://github.com/rustwasm/wasm-bindgen/issues/111