Fix uncontrolled component warning
Setting nature filters and then changing mons would log the following
console warning:

> A component is changing a controlled input of type checkbox to be uncontrolled.

Some quick [Googling](https://stackoverflow.com/questions/54364872/a-component-is-changing-an-uncontrolled-input-of-type-checkbox-to-be-controlled)
shows this happens when an input's value is `undefined`. Indeed, this
occurs when the natures filter is reset, since each checkbox has a value
of `checked` set to

```
value && value[nature]
```

so we can fix this by changing to

```
value ? value[nature] : false
```

which returns `false` when `value` is undefined.